reorganize atlab helper to allow it to be run as a module
