+++ neuropixel_pipeline/helpers/atlab/__init__.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+PROBE_CALIBRATION_DIR = Path("/mnt/lab/Neuropixel Electrode Calib Files")
+PROBE_CALIBRATION_SUFFIX = "_ADCCalibration.csv"
+DEFAULT_PROBE_TYPE = "neuropixels 1.0 - 3B"
+DEFAULT_PROBE_COMMENT = "Filled using calibration files"
+
